+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Whenever `minor` is bumped in the `major.minor.patch` version (e.g. `1.0.2 -> 1.
   - Support for custom special tokens, like `[missing]` or `[instruction fault]`
   - Various fixes for single-stream and dynamic assignment
   - Add `data_random` and `data_random_prob` for random item sampling that can be used for spot attention checks
+  - Dashboard now asks where to save `progress.json` and `annotations.json`.
 - v1.1.4
   - Make span toolboxes stick for 0.5 second after selection
 - v1.1.3